updated Jenkinsfile

+++ Jenkinsfile
@@ -2,7 +2,6 @@ pipeline {
     agent any
 
     environment {
-        KUBECONFIG = credentials('kubeconfig-file-id')
         APP_DIR = 'test/React/my-react-app'
         K8S_NAMESPACE = 'your-namespace'
         DEPLOYMENT_FILES = "test/selenium-hub.yaml,test/selenium-hub-service.yaml,test/selenium-node-chrome-deployment.yaml,test/selenium-node-edge-deployment.yaml,test/selenium-node-firefox-deployment.yaml"
@@ -58,8 +57,17 @@ pipeline {
         }
         always {
             script {
-                // Clean up by stopping the React development server
-                bat 'taskkill /F /IM node.exe'
+                try {
+                    // Clean up by stopping the React development server, check if node.exe is running
+                    def processList = bat(script: 'tasklist /FI "IMAGENAME eq node.exe"', returnStdout: true).trim()
+                    if (processList.contains('node.exe')) {
+                        bat 'taskkill /F /IM node.exe'
+                    } else {
+                        echo 'No Node.js process found to kill.'
+                    }
+                } catch (Exception e) {
+                    echo "Error during cleanup: ${e.getMessage()}"
+                }
             }
         }
     }
